refactored entire vulkanBackend so that #include<vulkan/vulkan.h> only appears in .cpp files! Texture and Buffer interface coupling can not continue!

+++ engine/interfaces/renderer/textureFormat.h
@@ -6,7 +6,7 @@
 
 namespace emberEngine
 {
-    enum class TextureFormatType : uint16_t
+    enum class TextureFormatType : uint32_t
     {
         // Fallback:
 		unknown = 0,        // unknown / unsupported format
